feat: breaks get colors out into a helper
