+++ src/modules/colorTable.js
@@ -1,5 +1,5 @@
-import Color from "colorjs.io";
 import slugify from '@sindresorhus/slugify';
+import getColors from './getColors.js'
 const contrastText = (contrast) => {
     const good = contrast >= 4.5;
     const goodForLargeText = contrast >= 3;
@@ -29,33 +29,8 @@ const contrastText = (contrast) => {
  */
 const colorList = (colors, headingLevel = 2, prefix = 'color-list') => {
 
-    const colorObjects = colors.map(color => {
-        try {
-            const colorObject = new Color(color.trim());
-            const colorName = color.name || color;
-            return {
-                name: colorName.trim(),
-                data: colorObject
-            }
-        } catch {
-            return false;
-        }
-    }).filter(Boolean)
-    colorObjects.forEach((color1,index,) => {
-        if(color1) {
-            colorObjects.forEach((color2) =>{
-                if(color1 !== color2){
-                    if(! colorObjects[index].colors) {
-                        colorObjects[index].colors = [];
-                    }
-                    if(color2) {
-                        const contrast = color1.data.contrast(color2.data,"WCAG21")
-                        colorObjects[index].colors.push({...color2, contrast: contrast})
-                    }
-                }
-            })
-        }
-    });
+    const colorObjects = getColors(colors);
+
     const heading = (text,h = headingLevel) => {
         const id = slugify(text)
         return `<h${h} id="${id}">
